Added localStorage
Added a local Storage so when ever the page refreshes the data will not be
lost. The state is preserved everytime a new task is added, updated or deleted
for all this type of functions we save it in the taskManagerTasks.
We also added init new function that reloads the tasks from local storage to
the tasks array.